+++ projects/kit/components/arrow/index.ts
@@ -1,3 +1,4 @@
 export * from './arrow.component';
 export * from './arrow.module';
 export * from './arrow.options';
+export * from './arrow.providers';
